requirements.txt
# Julius Baer AML Monitoring System - Python Dependencies
# 
# SYSTEM REQUIREMENTS (install before pip install):
# - macOS: brew install libomp tesseract
# - Linux: sudo apt install libgomp1 tesseract-ocr
# - Windows: Install Visual C++ Redistributable + Tesseract OCR
#
# For automated setup, run: ./setup.sh (macOS/Linux) or setup.bat (Windows)

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Data Processing & Analysis
pandas==2.1.3
numpy==1.24.3

# Machine Learning - EXACT VERSIONS TO PREVENT COMPATIBILITY ISSUES
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# AI & NLP
openai==1.3.5
groq==0.33.0
langchain==0.0.339
langchain-openai==0.0.2
sentence-transformers==2.2.2
chromadb==0.4.18

# Web & API
requests==2.31.0
beautifulsoup4==4.12.2

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
watchdog==3.0.0
psutil==5.9.6

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1